Fill in the test for rounded rectangle splash and highlights
Fixes https://github.com/flutter/flutter/issues/9031

Also some updates to mock_canvas that were needed to do this.
